Split gestureHandlers.ts into separate files for every handler
