+++ src/index.ts
@@ -9,31 +9,42 @@ export type {
   // event payloads types
   GestureEventPayload,
   HandlerStateChangeEventPayload,
+} from './handlers/gestureHandlerCommon';
+export type {
   TapGestureHandlerEventPayload,
-  ForceTouchGestureHandlerEventPayload,
-  LongPressGestureHandlerEventPayload,
-  PanGestureHandlerEventPayload,
-  PinchGestureHandlerEventPayload,
-  RotationGestureHandlerEventPayload,
-  FlingGestureHandlerEventPayload,
-  // gesture handlers props types
   TapGestureHandlerProps,
+} from './handlers/TapGestureHandler';
+export type {
+  ForceTouchGestureHandlerEventPayload,
   ForceTouchGestureHandlerProps,
+} from './handlers/ForceTouchGestureHandler';
+export type {
+  LongPressGestureHandlerEventPayload,
   LongPressGestureHandlerProps,
+} from './handlers/LongPressGestureHandler';
+export type {
+  PanGestureHandlerEventPayload,
   PanGestureHandlerProps,
+} from './handlers/PanGestureHandler';
+export type {
+  PinchGestureHandlerEventPayload,
   PinchGestureHandlerProps,
+} from './handlers/PinchGestureHandler';
+export type {
+  RotationGestureHandlerEventPayload,
   RotationGestureHandlerProps,
+} from './handlers/RotationGestureHandler';
+export type {
+  FlingGestureHandlerEventPayload,
   FlingGestureHandlerProps,
-} from './handlers/gestureHandlerCommon';
-export {
-  TapGestureHandler,
-  ForceTouchGestureHandler,
-  LongPressGestureHandler,
-  PanGestureHandler,
-  PinchGestureHandler,
-  RotationGestureHandler,
-  FlingGestureHandler,
-} from './handlers/gestureHandlerCommon';
+} from './handlers/FlingGestureHandler';
+export { TapGestureHandler } from './handlers/TapGestureHandler';
+export { ForceTouchGestureHandler } from './handlers/ForceTouchGestureHandler';
+export { LongPressGestureHandler } from './handlers/LongPressGestureHandler';
+export { PanGestureHandler } from './handlers/PanGestureHandler';
+export { PinchGestureHandler } from './handlers/PinchGestureHandler';
+export { RotationGestureHandler } from './handlers/RotationGestureHandler';
+export { FlingGestureHandler } from './handlers/FlingGestureHandler';
 export { default as createNativeWrapper } from './handlers/createNativeWrapper';
 export type {
   NativeViewGestureHandlerPayload,
